backend/requirements.txt
flask-wtf
flask-sqlalchemy
flask-migrate
flask-cors
# flask_praetorian
flask-login
flask-jwt-extended
cffi
flask-bcrypt